Fix Simulation build errors and add consistency checker (WIP: type mismatches to resolve)

+++ src/Host/ZakYip.WheelDiverterSorter.Host/Services/CachedSwitchingPathGenerator.cs
@@ -23,7 +23,7 @@ public class CachedSwitchingPathGenerator : ISwitchingPathGenerator {
         _logger = logger;
     }
 
-    public SwitchingPath? GeneratePath(int targetChuteId)
+    public SwitchingPath? GeneratePath(long targetChuteId)
     {
         if (targetChuteId <= 0)
         {
